requirements.txt
# === Base Dependencies ===
certifi==2025.10.5
charset-normalizer==3.4.3
filelock==3.20.0
fsspec==2025.9.0
idna==3.10
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
networkx==3.5
numpy==2.3.3
regex==2025.9.18
requests==2.32.5
setuptools==80.9.0
sympy==1.14.0
tiktoken==0.12.0
torch==2.8.0
typing_extensions==4.15.0
urllib3==2.5.0
wheel==0.45.1

# === Web / App ===
flask==2.3.3

# === RAG & AI Dependencies ===
python-dotenv==1.0.1
markdown==3.9
beautifulsoup4==4.12.2
pypdf==3.16.0
tqdm==4.65.0

# === LangChain & LLM Backends ===
langchain==0.3.27
langchain-groq==0.3.8
langchain-openai==0.3.35
langchain-huggingface==0.3.1
langchain-chroma==0.1.8
sentence-transformers==2.2.2
huggingface-hub==0.17.4
chromadb==0.4.3

